requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
anyio==3.6.1
    # via httpcore
attrs==22.1.0
    # via pytest
build==0.8.0
    # via pip-tools
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via
    #   pip-tools
    #   typer
commonmark==0.9.1
    # via rich
coverage[toml]==6.4.4
    # via pytest-cov
h11==0.12.0
    # via httpcore
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via
    #   -r requirements.in
    #   pytest-httpx
idna==3.3
    # via
    #   anyio
    #   rfc3986
iniconfig==1.1.1
    # via pytest
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   render-engine
markdown-it-py==2.1.0
    # via -r requirements.in
markdown2==2.4.3
    # via render-engine
markupsafe==2.1.1
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==8.14.0
    # via render-engine
packaging==21.3
    # via
    #   build
    #   pytest
pendulum==2.1.2
    # via render-engine
pep517==0.13.0
    # via build
pip-tools==6.8.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pygments==2.13.0
    # via rich
pyparsing==3.0.9
    # via packaging
pytailwindcss==0.1.4
    # via -r requirements.in
pytest==7.1.3
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-httpx
pytest-cov==3.0.0
    # via -r requirements.in
pytest-httpx==0.21.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via pendulum
python-frontmatter==1.0.0
    # via render-engine
python-slugify==6.1.2
    # via render-engine
pytzdata==2020.1
    # via pendulum
pyyaml==6.0
    # via python-frontmatter
render-engine==2022.8.1
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
rich==12.5.1
    # via render-engine
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   build
    #   coverage
    #   pep517
    #   pytest
typer==0.6.1
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
